[level 0] Title: 대문자와 소문자, Time: 17.82 ms, Memory: 79.1 MB -BaekjoonHub

+++ 프로그래머스/lv0/120893. 대문자와 소문자/대문자와 소문자.java
@@ -8,10 +8,10 @@ class Solution {
 		}
 		
 		for(int i = 0; i <arr.length; i++) {
-			if(65 <= (int)arr[i] && (int)arr[i] <= 90) {
+			if(65 <= arr[i] && arr[i] <= 90) {
 				answer += (char)((arr[i]) + 32);
 			}
-			else if(97 <= (int)arr[i] && (int)arr[i] <= 122) {
+			else if(97 <= arr[i] && arr[i] <= 122) {
 				answer += (char)((arr[i]) - 32);
 			}
 		}
